Prevents legacy path from mutating args
So much legacy code will (meaningfully or not) mutate the args provided
to it that it's safer to cache the Marshal.dump & #inspect of it as soon
as the Suture is created and then to reuse those. Oddly enough we don't
need the real args again.

Goal here is to limit our temptation to hold a reference to something 
that might get changed where it's easy for us to do so.

Fixes #19