Implement a call_limit option.
This may be useful if you have TONS of recordings and just want to 
verify a smattering of them at a time instead of sitting around all 
day watching slow tests.

This feature should be used _sparingly_, since `Suture.verify` tests are
not intended to be kept around forever, and the recordings _not tested_
at some point aren't able to provide any assurance or value. As a 
result, I'd recommend using this feature (and the forthcoming 
time_limit) in conjunction with code coverage analysis of the code paths

Fixes #7