If new & old are defined new should run by default
This was an oversight up until now. If Suture's not doing anything
special (like recording, call_both, fallback), it should call the new
path if it's defined.

The only exception to this is if SUTURE_DISABLE is set. In that case,
one assumes they want to hit the brakes on all new code and revert to 
the old stuff.

We should maybe remember to document this.