Create the seam
This is a no-op, just testing to make sure it invokes correctly.

A bit of explanation on the design of this seam:

The `:old` lambda is designed to "broaden the seam" to the entire state
that the function operates over, which in this case is essentially "item
in, mutated item out", and as such takes the item to be mutated as an
argument, and then returns it as the result of the lambda expression.