Start recording calls
With suture, we record calls so that we can—as we refactor or 
reimplement a code path—compare its past and future behavior in an
automated fashion (as a form of characterization testing).

I've unignored the db/suture.sqlite3 for the sake of example here after
capturing a few calls to a set of items I created through the web 
interface.