Write a test to validate recorded calls
Run the test with:

    bundle exec rake suture

This was a little involved and my first-ever attempt using Suture with
ActiveRecord inputs & outputs. A few notes.

* The rake task is completely divorced from the Rails app & the built-in
Rails testing infrastructure/tasks. This seems like a good idea since
it'll probably often be the case that users will depend on a database
snapshot (or, at the very least, different database setup than the 
fixtures/factories used in the rest of their test suite)

* The test requires the Rails app as well as the item model

* The test setup has to establish an AR connection on its
own in UpdateQualityTest#setup, which is a little awkward

* The test's `:subject` lambda is intentionally identical to the 
one from ItemsController, so that we can validate that the recordings
can be replicated in a controlled automated environment just like they
were when we populated db/suture.sqlite3 via exploratory testing

* Because ActiveRecord compares items by id, the default comparator is
insufficient at least until #30 is resolved