The Auditor now makes sure errors match, too
The happy side effect of this is that if the new code path raises
and the old code path does not raise (e.g. it returns happily),
and the user has also toggled to return the old value, it will
gracefully recover for user's sake if the user has also configured
to not raise mismatch errors.

Fixes #64